Use u32 instead of u64 for output index.
Indexing via tx.siacoin_outputs[vout_index] may be a common pattern. WASM usize is u32 which would require doing lossy u64->u32

+++ src/types/transaction.rs
@@ -391,11 +391,11 @@ pub type TransactionId = Hash256;
 pub struct SiacoinOutputId(pub Hash256);
 
 impl SiacoinOutputId {
-    pub fn new(txid: TransactionId, index: u64) -> Self {
+    pub fn new(txid: TransactionId, index: u32) -> Self {
         let mut encoder = Encoder::default();
         encoder.write_distinguisher("id/siacoinoutput");
         txid.encode(&mut encoder);
-        encoder.write_u64(index);
+        encoder.write_u64(index as u64);
         SiacoinOutputId(encoder.hash())
     }
 }
